pyproject.toml
[tool.poetry]
name = "joejoinerrbot"
version = "0.1.0"
description = ""
authors = ["Joe Joiner <joejoinerr@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10"
tweepy = "^4.10.0"
loguru = "^0.6.0"
transformers = "^4.20.0"
python-dotenv = "^0.20.0"
torch = "^1.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
